FIX check mandatory thirdparty fields for mass action

+++ htdocs/core/triggers/interface_80_modStripe_Stripe.class.php
@@ -35,46 +35,20 @@ require_once DOL_DOCUMENT_ROOT.'/core/triggers/dolibarrtriggers.class.php';
  */
 class InterfaceStripe extends DolibarrTriggers
 {
-    /**
-     * @var DoliDB Database handler.
-     */
-    public $db;
-
-    /**
-     *   Constructor
-     *
-     *   @param DoliDB $db Database handler
-     */
-    public function __construct($db)
-    {
-        $this->db = $db;
-
-        $this->name = preg_replace('/^Interface/i', '', get_class($this));
-	    $this->family = 'stripe';
-        $this->description = "Triggers of the module Stripe";
-        $this->version = 'dolibarr'; // 'development', 'experimental', 'dolibarr' or version
-        $this->picto = 'stripe';
-    }
-
 	/**
-	 * Trigger name
+	 *   Constructor
 	 *
-	 * @return string Name of trigger file
+	 *   @param DoliDB $db Database handler
 	 */
-	public function getName()
+	public function __construct($db)
 	{
-		return $this->name;
-	}
+		$this->db = $db;
 
-
-	/**
-	 * Trigger description
-	 *
-	 * @return string Description of trigger file
-	 */
-	public function getDesc()
-	{
-		return $this->description;
+		$this->name = preg_replace('/^Interface/i', '', get_class($this));
+		$this->family = 'stripe';
+		$this->description = "Triggers of the module Stripe";
+		$this->version = self::VERSION_DOLIBARR; // 'development', 'experimental', 'dolibarr' or version
+		$this->picto = 'stripe';
 	}
 
 	/**
@@ -104,8 +78,7 @@ class InterfaceStripe extends DolibarrTriggers
 
 		$service = 'StripeTest';
 		$servicestatus = 0;
-		if (!empty($conf->global->STRIPE_LIVE) && !GETPOST('forcesandbox', 'alpha'))
-		{
+		if (!empty($conf->global->STRIPE_LIVE) && !GETPOST('forcesandbox', 'alpha')) {
 			$service = 'StripeLive';
 			$servicestatus = 1;
 		}
@@ -118,8 +91,7 @@ class InterfaceStripe extends DolibarrTriggers
 
 			if ($object->client != 0) {
 				$customer = $stripe->customerStripe($object, $stripeacc, $servicestatus); // This make a network request
-				if ($customer)
-				{
+				if ($customer) {
 					$namecleaned = $object->name ? $object->name : null;
 					$vatcleaned = $object->tva_intra ? $object->tva_intra : null; // Example of valid numbers are 'FR12345678901' or 'FR12345678902'
 					$desccleaned = $object->name_alias ? $object->name_alias : null;
@@ -147,15 +119,13 @@ class InterfaceStripe extends DolibarrTriggers
 					if ($desccleaned != $customer->description) $changerequested++;
 					if (($customer->tax_exempt == 'exempt' && !$object->tva_assuj) || (!$customer->tax_exempt == 'exempt' && empty($object->tva_assuj))) $changerequested++;
 					if (!isset($customer->tax_ids['data']) && !is_null($vatcleaned)) $changerequested++;
-					elseif (isset($customer->tax_ids['data']))
-					{
+					elseif (isset($customer->tax_ids['data'])) {
 						$taxinfo = reset($customer->tax_ids['data']);
 						if (empty($taxinfo) && !empty($vatcleaned)) $changerequested++;
 						if (isset($taxinfo->value) && $vatcleaned != $taxinfo->value) $changerequested++;
 					}
 
-					if ($changerequested)
-					{
+					if ($changerequested) {
 						/*if (! empty($object->email)) $customer->email = $object->email;
 						$customer->description = $namecleaned;
 						if (empty($taxinfo)) $customer->tax_info = array('type'=>'vat', 'tax_id'=>null);
@@ -167,24 +137,17 @@ class InterfaceStripe extends DolibarrTriggers
 
 						try {
 							// Update Tax info on Stripe
-							if (!empty($conf->global->STRIPE_SAVE_TAX_IDS))	// We setup to save Tax info on Stripe side. Warning: This may result in error when saving customer
-							{
-								if (!empty($vatcleaned))
-								{
+							if (!empty($conf->global->STRIPE_SAVE_TAX_IDS)) {	// We setup to save Tax info on Stripe side. Warning: This may result in error when saving customer
+								if (!empty($vatcleaned)) {
 									$isineec = isInEEC($object);
-									if ($object->country_code && $isineec)
-									{
+									if ($object->country_code && $isineec) {
 										//$taxids = $customer->allTaxIds($customer->id);
 										$customer->createTaxId($customer->id, array('type'=>'eu_vat', 'value'=>$vatcleaned));
 									}
-								}
-								else
-								{
+								} else {
 									$taxids = $customer->allTaxIds($customer->id);
-									if (is_array($taxids->data))
-									{
-										foreach ($taxids->data as $taxidobj)
-										{
+									if (is_array($taxids->data)) {
+										foreach ($taxids->data as $taxidobj) {
 											$customer->deleteTaxId($customer->id, $taxidobj->id);
 										}
 									}
@@ -193,10 +156,8 @@ class InterfaceStripe extends DolibarrTriggers
 
 							// Update Customer on Stripe
 							$customer->save();
-						}
-						catch (Exception $e)
-						{
-						    //var_dump(\Stripe\Stripe::getApiVersion());
+						} catch (Exception $e) {
+							//var_dump(\Stripe\Stripe::getApiVersion());
 							$this->errors[] = $e->getMessage();
 							$ok = -1;
 						}
@@ -207,17 +168,17 @@ class InterfaceStripe extends DolibarrTriggers
 		if ($action == 'COMPANY_DELETE') {
 			dol_syslog("Trigger '".$this->name."' for action '$action' launched by ".__FILE__.". id=".$object->id);
 
-			$stripeacc = $stripe->getStripeAccount($service); // No need of network access for this. May return '' if no Oauth defined.
+			if (!empty($conf->global->STRIPE_DELETE_STRIPE_ACCOUNT_WHEN_DELETING_THIRPARTY)) {
+				// By default, we do not delete the stripe account. We may need to reuse it with its payment_intent, for example if delete is for a merge of thirdparties.
+				$stripeacc = $stripe->getStripeAccount($service); // No need of network access for this. May return '' if no Oauth defined.
 
-			$customer = $stripe->customerStripe($object, $stripeacc, $servicestatus);
-			if ($customer)
-			{
-				try {
-					$customer->delete();
-				}
-				catch (Exception $e)
-				{
-					dol_syslog("Failed to delete Stripe customer ".$e->getMessage(), LOG_WARNING);
+				$customer = $stripe->customerStripe($object, $stripeacc, $servicestatus);
+				if ($customer) {
+					try {
+						$customer->delete();
+					} catch (Exception $e) {
+						dol_syslog("Failed to delete Stripe customer ".$e->getMessage(), LOG_WARNING);
+					}
 				}
 			}
 
@@ -233,13 +194,11 @@ class InterfaceStripe extends DolibarrTriggers
 		if ($action == 'COMPANYPAYMENTMODE_MODIFY' && $object->type == 'card') {
 			dol_syslog("Trigger '".$this->name."' for action '$action' launched by ".__FILE__.". id=".$object->id);
 
-			if (!empty($object->stripe_card_ref))
-			{
+			if (!empty($object->stripe_card_ref)) {
 				$stripeacc = $stripe->getStripeAccount($service); // No need of network access for this. May return '' if no Oauth defined.
 				$stripecu = $stripe->getStripeCustomerAccount($object->fk_soc); // No need of network access for this
 
-				if ($stripecu)
-				{
+				if ($stripecu) {
 					// Get customer (required to get a card)
 					if (empty($stripeacc)) {				// If the Stripe connect account not set, we use common API usage
 						$customer = \Stripe\Customer::retrieve($stripecu);
@@ -247,16 +206,13 @@ class InterfaceStripe extends DolibarrTriggers
 						$customer = \Stripe\Customer::retrieve($stripecu, array("stripe_account" => $stripeacc));
 					}
 
-					if ($customer)
-					{
+					if ($customer) {
 						$card = $stripe->cardStripe($customer, $object, $stripeacc, $servicestatus);
 						if ($card) {
 							$card->metadata = array('dol_id'=>$object->id, 'dol_version'=>DOL_VERSION, 'dol_entity'=>$conf->entity, 'ipaddress'=>(empty($_SERVER['REMOTE_ADDR']) ? '' : $_SERVER['REMOTE_ADDR']));
 							try {
 								$card->save();
-							}
-							catch (Exception $e)
-							{
+							} catch (Exception $e) {
 								$ok = -1;
 								$this->error = $e->getMessages();
 							}
@@ -268,13 +224,11 @@ class InterfaceStripe extends DolibarrTriggers
 		if ($action == 'COMPANYPAYMENTMODE_DELETE' && $object->type == 'card') {
 			dol_syslog("Trigger '".$this->name."' for action '$action' launched by ".__FILE__.". id=".$object->id);
 
-			if (!empty($object->stripe_card_ref))
-			{
+			if (!empty($object->stripe_card_ref)) {
 				$stripeacc = $stripe->getStripeAccount($service); // No need of network access for this. May return '' if no Oauth defined.
 				$stripecu = $stripe->getStripeCustomerAccount($object->fk_soc); // No need of network access for this
 
-				if ($stripecu)
-				{
+				if ($stripecu) {
 					// Get customer (required to get a card)
 					if (empty($stripeacc)) {				// If the Stripe connect account not set, we use common API usage
 						$customer = \Stripe\Customer::retrieve($stripecu);
@@ -282,8 +236,7 @@ class InterfaceStripe extends DolibarrTriggers
 						$customer = \Stripe\Customer::retrieve($stripecu, array("stripe_account" => $stripeacc));
 					}
 
-					if ($customer)
-					{
+					if ($customer) {
 						$card = $stripe->cardStripe($customer, $object, $stripeacc, $servicestatus);
 						if ($card) {
 							if (method_exists($card, 'detach')) $card->detach();
